Refactor move ActionInterface
